added background color, ux_ui folder
