--- my-react-app/src/ButtonActivity.tsx
@@ -1,14 +1,0 @@
-export default function ButtonActivity({ idButton, choice, style }) {
-	return (
-		<>
-			<button
-				id={idButton}
-				style={{ boxShadow: `${style.box_shadow} ${style.color}` }}
-				className="button-activity"
-				type="button"
-			>
-				Je veux <span style={{ color: style.color }}>{choice}</span>
-			</button>
-		</>
-	);
-}
