+++ my-react-app/src/Pages/SectionLearn/SectionLearn.css
@@ -1,10 +1,14 @@
-/* body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 40em;
-  background-color: #FC03DF;
-} */
+.pageLearn {
+    background-color: var(--main-color-apprendre);
+    width: 100%;  
+    height: 100vh;
+    display: flex; 
+    flex-direction: column; 
+    align-items: center;
+    justify-content: center; 
+    margin: 0; 
+    padding: 0; 
+}
 
 .Api {
   width: 90%; 
@@ -41,9 +45,9 @@
   gap: 20px; 
 }
 
-.container {
+/* .container {
   display: flex;
   flex-direction: column; 
   align-items: center; 
   margin-top: 20px; 
-}
+} */
