+++ my-react-app/src/Pages/SectionMeme/SectionMeme.css
@@ -1,3 +1,16 @@
+.pageMeme {
+    background-color: var(--main-color-amuser);
+    width: 100%;  
+    height: 100vh;
+    display: flex; 
+    flex-direction: column; 
+    align-items: center;
+    justify-content: center; 
+    margin: 0; 
+    padding: 0; 
+}
+
+
 .meme {
     width: 500px;
     height: fit-content;
@@ -8,11 +21,12 @@
 }
 
 .container_buttons {
-    justify-content: space-between;
+    display: flex;
+    justify-content: center;
     align-items: center;
     margin-top: 20px;
     width: 100%;
     max-width: 500px;
     margin: 20px auto;
-
+    gap: 20px;
 }
