+++ my-react-app/src/Pages/SectionPoem/SectionPoem.css
@@ -1,3 +1,15 @@
+.pagePoem {
+    background-color: var(--main-color-inspirer);
+    width: 100%;  
+    height: 100vh;
+    display: flex; 
+    flex-direction: column; 
+    align-items: center;
+    justify-content: center; 
+    margin: 0; 
+    padding: 0; 
+}
+
 .container-poem {
     width: 100%;
     display: flex;
