+++ my-react-app/src/components/ButtonRecharge/ButtonRecharge.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import './ButtonRecharge.css';
 import { useState, useEffect } from 'react';
-import SectionMeme from '../../Pages/SectionMeme/SectionMeme';
 
-export default function ButtonRecharge({ onClick }) {
+
+export default function ButtonRecharge() {
     return (
     <div>
         <button className="button-recharge" onClick={()=> location.reload()}><img src='../assets/icon.png'/></button>
