+++ my-react-app/src/components/ButtonRetour/ButtonRetour.css
@@ -12,7 +12,6 @@
     font-family: var(--button-font-family);
     cursor: pointer;
     margin: auto;
-    margin-top: 20px;
     box-shadow: 3px 3px 5px var(--main-color-text-dark);
 }
