+++ my-react-app/src/components/Panneau/Panneau.css
@@ -5,7 +5,7 @@
     position: relative;       
     box-shadow:  5px 5px 6px rgba(0, 0, 0, 0.3);
     width: 90vw;
-    height: 95vh;
+    height: 70vh;
     overflow-y: scroll;
     padding: 3rem 1.5rem;
     margin: 2rem;
