+++ my-react-app/src/index.css
@@ -7,7 +7,7 @@
 
     --main-color-amuser: #FED440;
     --main-color-inspirer: #49CEB9;
-    --main-color-apprendre: #FFC3A9;
+    --main-color-apprendre: #FC03DF;
     --main-color-bg-button: #B9BBBB;
 
     --main-color-text-light: #EEF1F0;
